Setup the brightness level successfully

+++ app/src/main/java/com/bitmavrick/lumolight/ui/tab/screenFlash/ScreenFlashViewModel.kt
@@ -27,7 +27,7 @@ class ScreenFlashViewModel : ViewModel() {
         }
     }
 
-    fun updateScreenFlashBrightness(index: Int, brightness: Int){
+    fun updateScreenFlashBrightness(index: Int, brightness: Float){
         _uiState.update {
             it.copy(
                 screenFlashBrightnessIndex = index,
